// Admin/src/app/core/models/patient.models.ts
export interface patient {
  clinicname: any;
  clinicid: string;
  mpino: string;
  oldfileno: string;
  fileno: string;
  title: string;
  agep: string;
  guardian_mpino: string;
  firstname: string;
  lastname: string;
  fullname: string;
  idnumber: string;
  dob: string;
  passportno: string;
  expdt: string;
  gender: string;
  race: string;
  language: string;
  maritalstatus: string;
  citizenship: string;
  religion: string;
  patient_category: string;
  bill_address: string;
  bill_suburd: string;
  bill_city: string;
  bill_areacode: string;
  bill_region: string;
  bill_ward: string;
  patientpno: string;
  nationality: string;
  patientclassification: string;
  patientemail: string;
  patientextra: string;
  kin_detail: string;
  status: string;
  addate: string;
  addtime: string;
  appointmentsid: number;
  regstart: string;
  regend: string;
  chld_name: string;
  chld_age: number;
  chld_dob: string;
  chld_weight: number;
  child_id: number;
  patient_tokenid: number;
  patient_token_start: string;
  patient_token_end: string;
  medicalfund: string;
  medicalfundname: string;
  medicalfundno: string;
  pempstatus: string;
  pemployer: string;
  pempaddress: string;
  pempcountry: string;
  pempprovince: string;
  pempsuburb: string;
  pempcity: string;
  pempcode: string;
  pempcontact: string;
  pempemail: string;
  patientbarcode: string;
  ptype: string;
  left_thumb_template_base64: string;
  left_thumb: string;
  left_index_template_base64: string;
  left_index: string;
  right_thumb_template_base64: string;
  right_thumb: string;
  right_index_template_base64: string;
  right_index: string;
  pemployed: string;
  pnextofkin: string;
  patientdependant: string;
  added_by: string;
  added_userid: string;
}

export interface PatientResponse {
  message: string;
  visits: any;
  classifications: any;
  patients: patient[];
}